fixed several code smells
