+++ src/Mealz/MealBundle/Event/MealOfferCancelledEvent.php
@@ -20,7 +20,7 @@ class MealOfferCancelledEvent extends Event
         $this->participant = $participant;
     }
 
-    public function getMeal(): Meal
+    public function getMeal(): ?Meal
     {
         return $this->participant->getMeal();
     }
